refactor(web): decompose Sidebar into presentational sub-navs
Extract the workspace/build-units/teams navs and the create-team modal. All
useLiveQuery subscriptions stay in Sidebar (it is the always-mounted subscriber
that keeps the spine collections warm per ARCHITECTURE.md §6) — the sub-navs only
receive the queried data as props, so GC behavior is unchanged. Sidebar drops
419->206 lines; ~20 fewer typecheck errors from typing the prop boundaries.
Behavior-preserving.
